LayoutTests/platform/mac/fast/text/trailing-word-parse-expected.txt

PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
PASS declaration.getPropertyValue('-apple-trailing-word') is ""
PASS computedStyle.getPropertyCSSValue('-apple-trailing-word').cssText is "auto"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 1
PASS declaration.getPropertyValue('-apple-trailing-word') is "initial"
PASS computedStyle.getPropertyCSSValue('-apple-trailing-word').cssText is "auto"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 1
PASS declaration.getPropertyValue('-apple-trailing-word') is "-webkit-partially-balanced"
PASS computedStyle.getPropertyCSSValue('-apple-trailing-word').cssText is "-webkit-partially-balanced"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
PASS computedStyle.getPropertyCSSValue('-apple-trailing-word').cssText is "auto"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
PASS computedStyle.getPropertyCSSValue('-apple-trailing-word').cssText is "auto"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
PASS computedStyle.getPropertyCSSValue('-apple-trailing-word').cssText is "auto"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
PASS computedStyle.getPropertyCSSValue('-apple-trailing-word').cssText is "auto"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
PASS computedStyle.getPropertyCSSValue('-apple-trailing-word').cssText is "auto"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
PASS computedStyle.getPropertyCSSValue('-apple-trailing-word').cssText is "auto"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 1
PASS declaration.getPropertyValue('-apple-trailing-word') is "auto"
PASS computedStyle.getPropertyCSSValue('-apple-trailing-word').cssText is "auto"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
PASS computedStyle.getPropertyCSSValue('-apple-trailing-word').cssText is "auto"
PASS successfullyParsed is true

TEST COMPLETE
